chore: rename package to com.akrafts.addon
