+++ composeApp/src/androidMain/kotlin/com/akrafts/addon/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.kmpapp
+package com.akrafts.addon
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
