+++ composeApp/src/commonMain/kotlin/com/akrafts/addon/data/MuseumApi.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.kmpapp.data
+package com.akrafts.addon.data
 
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
