+++ composeApp/src/commonMain/kotlin/com/akrafts/addon/data/MuseumRepository.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.kmpapp.data
+package com.akrafts.addon.data
 
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.SupervisorJob
