+++ composeApp/src/commonMain/kotlin/com/akrafts/addon/data/MuseumStorage.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.kmpapp.data
+package com.akrafts.addon.data
 
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
