+++ composeApp/src/commonMain/kotlin/com/akrafts/addon/screens/EmptyScreenContent.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.kmpapp.screens
+package com.akrafts.addon.screens
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.material3.Text
