+++ composeApp/src/commonMain/kotlin/com/akrafts/addon/screens/detail/DetailViewModel.kt
@@ -1,8 +1,8 @@
-package com.jetbrains.kmpapp.screens.detail
+package com.akrafts.addon.screens.detail
 
 import androidx.lifecycle.ViewModel
-import com.jetbrains.kmpapp.data.MuseumObject
-import com.jetbrains.kmpapp.data.MuseumRepository
+import com.akrafts.addon.data.MuseumObject
+import com.akrafts.addon.data.MuseumRepository
 import kotlinx.coroutines.flow.Flow
 
 class DetailViewModel(private val museumRepository: MuseumRepository) : ViewModel() {
