+++ composeApp/src/commonMain/kotlin/com/akrafts/addon/screens/list/ListScreen.kt
@@ -1,4 +1,4 @@
-package com.jetbrains.kmpapp.screens.list
+package com.akrafts.addon.screens.list
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.foundation.background
@@ -26,8 +26,8 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import coil3.compose.AsyncImage
-import com.jetbrains.kmpapp.data.MuseumObject
-import com.jetbrains.kmpapp.screens.EmptyScreenContent
+import com.akrafts.addon.data.MuseumObject
+import com.akrafts.addon.screens.EmptyScreenContent
 import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
