+++ composeApp/src/commonMain/kotlin/com/akrafts/addon/screens/list/ListViewModel.kt
@@ -1,9 +1,9 @@
-package com.jetbrains.kmpapp.screens.list
+package com.akrafts.addon.screens.list
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.jetbrains.kmpapp.data.MuseumObject
-import com.jetbrains.kmpapp.data.MuseumRepository
+import com.akrafts.addon.data.MuseumObject
+import com.akrafts.addon.data.MuseumRepository
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.stateIn
