+++ iosApp/Configuration/Config.xcconfig
@@ -1,3 +1,3 @@
 TEAM_ID=
-BUNDLE_ID=com.jetbrains.kmpapp.KMP-App-Template
+BUNDLE_ID=com.akrafts.addon.KMP-App-Template
 APP_NAME=KMP App
